requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cloudinary==1.4.0
decorator==4.0.10
Django==1.8
django-bower==5.1.0
django-envie==0.0.9
django-webpack-loader==0.3.3
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
enum34==1.1.6
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
PyYAML==3.12
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.0
wcwidth==0.1.7
